server/requirements.txt
Django>=3.1.1
djangorestframework>=3.12.1
psycopg2-binary>=2.8.6
djangorestframework-simplejwt>=4.5.0
django-cors-headers>=3.6.0